introduce a first draft of a factory for analysis configuration blocks
This file defines a factory method that can create a configuration
block sequence based on a passed in name.  This avoids having to
import all the various config block sequence makers in the
configuration code, and also would make it easier to create them
from a text configuration file.

This relies heavily on the blocks exposing all configurable
parameters as options, since there is no other mechanism to
configure them through this interface.

The implementation itself is probably not the best possible, it
lacks all extensibility, gathers all information in a single place,
etc.  Still for now this ought to be good enough.