add options to the remaining analysis configuration blocks
This tries to give the same treatment given to muons and photons to
all the other configurations, so that it should be possible to
configure all the blocks just through the options.

There is probably some cleanup that could be done here in the future,
regarding what some of the configuration blocks should be called
exactly, what the options exposed should be etc.  The main goal of
this update is that I can now (more or less awkwardly) set everything
I can via options that was previously limited to named parameters to
the various make...Config() calls.

I also added customizable behavior for setting a value of `None` on an
option.