AthContainersRoot+xAODAccess: Dictionary fix.
AthContainersRoot: Dictionary fix.

When we	look up	the name of a factory class, we	need to	try to match
exactly	the name in TClassTable.  That now means including the 
allocator template argument, and also stripping	std:: prefixes.
Otherwise, as can trigger autoparsing, which can then trip over
ROOT bugs in the dbg build.