Adding an AMI Tag helper for CA configuration
Adding an AMI Tag helper for CA configuration

This adds a CA-based helper for getting and setting the AMI tag in the
tag info metadata. The output works for Derivation_tf, using the
--AMITag from the runArgs.

The metadata collection from the input file isn't working yet, because
the MD helper doesn't seem to access the TagInfo metadata. Suggestions
there are welcome!