Restoring the CLI options and switches to visualize the AFP geo.
AFP had been added to VP1 in Athena rel.20. However, when migrating from
SVN to Git, an earlier version was picked and that particular tag was
left behind, in SVN.

This is the SVN version, where AFP was IN:

https://svn-atlas.web.cern.ch/websvn/filedetails.php?repname=repository&path=%2Fgraphics%2FVP1%2FVP1Systems%2FVP1GeometrySystems%2Ftags%2FVP1GeometrySystems-00-16-11%2FVP1GeometrySystems%2FVP1GeoFlags.h

And this is the version in current Git 'master' branch, where AFP is not
present:

https://gitlab.cern.ch:8443/atlas/athena/-/blob/master/graphics/VP1/VP1Systems/VP1GeometrySystems/VP1GeometrySystems/VP1GeoFlags.h

This commit, and the following ones, will restore in Athena 'master'
the possibility of visualizing the AFP geometry in VP1.