TriggerJobOpts: fix HLT trigger content for Run1 AOD, RAWtoALL workflow, non-CA implementation
Hello,

This MR is fixing the HLT content for Run1 when running RAWtoALL and the non-CA implementation (HLTTriggerResultGetter).
Currently this workflow crashes (ATR-26630) because we are trying to write old-type objects (Run1 EDM) no longer supported.
* When writing out ESDs, we write HLT containers after xAOD conversion (`edmVersion` is forced to 2 i.e. Run2 EDM when running over a Run1 input):
https://gitlab.cern.ch/atlas/athena/-/blob/master/Trigger/TriggerCommon/TriggerJobOpts/python/HLTTriggerResultGetter.py#L63-66
* But for the AOD, we write the original EDM version (`flags.Trigger.EDMVersion` = 1 for Run1).

With this MR, we force the HLT Run2 EDM for the AOD as well.

The problem was revealed when we moved the Run1 trigger ART test to RAWtoALL. Before, we were running RAWtoESD then ESDtoAOD, and in the ESDtoAOD step the problem was hidden because `flags.Trigger.EDMVersion` is 2 (the ESD contains the Run2 EDM).

For the CA implementation, we should have no such problem:
https://gitlab.cern.ch/atlas/athena/-/blob/master/Trigger/TriggerCommon/TriggerJobOpts/python/TriggerRecoConfig.py#L328-330

Cheers,
Bertrand