Set the pt and chi2min limits for the tau and bjet tracking instances
Set the pt and chi2min limits for the tau and bjet tracking instances

The ptmin for the bjet instance was previously set, this request also adds the limit
for the tau instances. It also sets the chi2min for the tauIso and bjet tracking to 12
and can be over ridden by a preexec.

In addition it adds the Roi z width to the tracking configuration, and takes the settings
used for the slice from the configuration rather than using the default parameters - the
settings from the slice config currently have the same values as the default so the current
settings of the Roi widths should not change.