GeneratorObjectsTPCnv: Fix DataPool usage for HepMC3.
With HepMC3, when we make a shared_ptr from an object we're just allocated
from a DataPool, we need to ensure that there are no other shared_ptr's
referencing that object; otherwise the behavior is undefined (but can
manifest as the weak references in GenParticle mysteriously expiring;
see ATR-26790).

Set up clear hooks on the DataPools so that the objects get cleared
when they're returned to the pools.