tauRec+tauRecTools: implement new Deep Set NN tau ID
tauRec+tauRecTools: unleash new Deep Set NN tau ID!

Hello,

This MR is deploying the improved tau ID based on a deep set NN, which is meant to deplace the current default RNN tau ID.
It is largely reusing the RNN implementation. The NN uses a bunch of extra variables compared to the RNN, namely the RNN track classification scores (from the RNN tau track classifier, 4 probabilities per track).
Support has been added for both RecExCommon and CA configurations.

The calibration files (3 json files for the NN and 3 ROOT files for the WPs) are on EOS and will soon appear on cvmfs.

The goal is to run this new tau ID in RAWtoALL in R23.
For R22, we'll deploy it in AODtoDAOD via a wrapper in DerivationFrameworkTau, in a MR yet to come, hopefully tomorrow, so that the deep set ID goes into the next DAOD_PHYS production to be launched imminently.

Cheers,
Bertrand