Clean-up of legacy config in derivation framework: DerivationFrameworkMCTruth
Clean-up of legacy config in derivation framework: DerivationFrameworkMCTruth

This is a continuation of the clean-up of obsolete legacy config files in the derivation framework. This MR targets DerivationFrameworkMCTruth. Since common legacy config files are removed, this now breaks RecExCommon-based DAOD jobs in master (at least for MC). Consequently it may break some ART tests. Since we are now using CA for production, this is desirable since we need to move such ART tests to CA so they reflect what is being run.

Once this is merged and we have cleaned up any consequent crashes in ART, the larger part of the clean-up will be done.