DerivationFrameworkPHYS: fix tau thinning in PHYSLITE and add DeepSet tau ID to PHYSLITE output
DerivationFrameworkPHYS: fix tau thinning in PHYSLITE and add DeepSet tau ID to PHYSLITE output

Hello,

This MR fixes a problem with tau thinning in PHYSLITE caused by !59677 : the TauTrackParticleThinning was dropped from the list of thinning tools supported in PhysCommonThinningConfig and replaced with an all-in-one tau thinning tool; this was propagated to the PHYS configuration but not to PHYSLITE.

This MR drops the TauTrackParticleThinning from PHYSLITE and only keeps the new TauThinningTool.
It fixes all the PHYSLITE ART tests that are failing today.

Besides, I have added the new Deep Set tau ID to the PHYSLITE output, that's 2 floats (scores) and 4 chars (WPs).
The new tau ID was introduced in !59725 but the output was not kept in PHYSLITE.

For completeness, there is no tau thinning issue in PHYS->PHYSLITE, and the PHYSLITE output seems to be the same (for what concerns taus) when running AOD->PHYSLITE and AOD->PHYS->PHYSLITE.

Unrelated to this MR, I also checked that the multiple tau MRs from yesterday didn't introduce differences in other objects for the PHYS format (I compared today's nightly with yesterday's nightly). I had to ignore tau branches and InDetTrackParticles for this check, as these are affected by tau thinning.

Cheers,
Bertrand