tauMonitoring+TauDQA: compute tau track "rnn_conversionScore" variable from unitarity
tauMonitoring+TauDQA: compute tau track "rnn_conversionScore" variable from unitarity

Hello,

This MR is fixing a crash in the tau PhysVal monitoring (ATLSWUPGR-178) that appeared as a consequence of !59677.
We now provide 3 new variables per tau track in DAODs (`rnn_chargedScore`, `rnn_isolationScore`, `rnn_conversionScore`), which correspond to 3 probabilities, and there exists a 4th probability (`rnn_fakeScore`) computed in RAWtoALL but it's not written out as it can be obtained from unitarity, the 4 probabilities add up to 1.
However, in the PhysVal (and offline DQ) tau monitoring, we were assuming that if `rnn_chargedScore` was available, then all 4 probabilities were available.

This MR instead computes `rnn_fakeScore` from unitarity. It is necessary for PhysVal, as the tau variables are coming from the tau smart slimming list that only provides 3 probabilities.
I also implemented this for the offline tau DQ monitoring, but I'm not sure it is necessary (if the monitoring is run in RAWtoALL or on AODs, the 4 probabilities are available).

Cheers,
Bertrand