STG YvsX on track plots
STG YvsX on track plots:

Cluster on track implemented. The information from cluster is used to retrieve the position ath which the track intersects
with the STG active surface. This allow to retrieve the X and Y coordinates of the track and to perform YvsX global hit maps
per channel.