Introduced new package: AtlasGeometryCommon/CrackRegionGeoModel
The reason for introducing this package is to have a way of building
dead material description either from the LAr GeoModel Tool (old way)
or from the Service Extension Tool (new way)

CrackRegionGeoModel currently contains code which used to live inside
BarrelDMConstruction class. As a result of this migration the latter
class became obsolete, and it was dropped.