tauRec: define ConfigFlags for tau output containers
tauRec: define ConfigFlags for tau output containers

Hello,

This MR is reworking the ConfigFlags used in the tau reconstruction with CA.
* We define config flags for container names, to avoid hardcoded names in multiple places.
The container names are now passed from one algorithm to the next via flags.
* The main goal of this MR is to define two flag categories, one for standard tau reconstruction ("Tau.TauRec"), one for the boosted electron-tau reconstruction (""Tau.TauEleRM"") that is being implemented in !57968. This "electron-subtracted tau reconstruction" is almost identical to the standard tau reconstruction, so these two types of reconstruction are becoming two passes of a common reconstruction setup. The implementation is based on the tracking configuration and its multiple passes. The flags are switched from standard to electron-subtracted by a cloneAndReplace operation, with "Tau.ActiveConfig" representing the config actively in use in a given pass.

Please note that the electron-subtracted tau configuration pieces are currently placeholders to illustrate how things will be organised. This part is not thoroughly validated, and the relevant tools and algorithms are not yet implemented (!57968). The electron-subtracted tau reconstruction is switched off by default (flags.Tau.doTauEleRMRec=False).
I have checked that the tau reconstruction output is unchanged in RAWtoALL with CA, using 500 data22 events.

Minor cleanup has also been made for the configuration of some tools, i.e. TauCellVariables (no need to configure parameters when default values are used) and TauCalibrateLC (the options that have been dropped are deprecated).

Cheers,
Bertrand