AthenaMPConfig: using single event dispatching if number of maxEvents is small.
Using single event dispatching if number of maxEvents is small.

This fixes the problem appearing in a very short jobs, when some workers do not get to process any events. In particular it means that FileMetaDataCreatorTool::postExecute may not be called and such jobs try to write FileMetaData from the input file.

Fixes ATEAM-882