reorganize and split the unit tests in AnalysisAlgorithmsConfig
The main thing this does is to split the comparison tests between
sequence and block configuration into three tests each to allow the
block and sequence jobs in parallel and then run the comparison
subsequently.  With this working I went ahead and then used the same
test structure in AnalysisBase as well.

This also required some minor fixes, as well as putting in some tests
in a disabled state, because logically they ought to exist, but they
are not working and were not present/disabled up to now.