Read/Write/Decor handles for TruthDecayCollectionMaker
Read/Write/Decor handles for TruthDecayCollectionMaker

This is the first of many MRs that will be needed to enable the derivation framework to run in AthenaMT, since many of the tools (especially in the MC truth domain) don't use R/W/D handles. To avoid a monster MR I'm going to migrate them one tool at a time.

Although this necessarily touches C++ it only impacts the reading/writing so will have no effect on the output of the tool. I have tested this on PHYS in serial but a test with MT will require all tools to be migrated.