EventLoopGrid: Change ELG_jediState.py return type to int to avoid binding problems
See JIRA ATLASG-2449 , using `TPython::Eval` with string as return type
causes issues with python3.7+. This is used in `PrunDriver.cxx`
to get JEDI task status for wait/retrieve. Changing type to int based on
enum defined both in `ELG_jediState.py` and `PrunDriver.cxx` to solve the issue.

Also cleaning-up some commented lines.

Note to return of the ELG_jediState:

Task states not mentioned above the state as a string used to be added
to nc_ELG_state_details meta information. I had a look and in whole athena
this is used in PrunDriver and only for print-outs so I think its okay
to just say "its not these states"?

Actually covering all the cases with enum would be messy
(there is around twenty possible states, see here:
    https://github.com/PanDAWMS/panda-bigmon-core/blob/master/core/constants.py#L85
)