AllConfigFlags: add Trigger.doLVL1 and doHLT in all releases
For releases without the Trigger flags, add `doLVL1 and `doHLT`
(defaulted to `False`) to make sure some of our CA code also works in
non-Gaudi environments. This e.g. allows to have code like
```
if flags.Trigger.doHLT and flags.Trigger.XYZ:
```
to protect the access to more advanced trigger flags that are not
available.