athenaHLT: remove usage of python include when running from DB/JSON
When running in athenaHLT from the DB or JSON files, we use
`TrigPSCPythonDbSetup.py` to setup a minimal Python environment. Change
this file from being an "include file" to a Python module.

Introduce helpers functions to alternatively include or import the file
based on its file name extension. This will be important to support
CA-based configurations.