Fix Acts Fatras Simulation
Even though the `ATLAS-R2-2016-00-00-00` is considered obsolete for release 23 onwards. It is OK
to use the ACTS material map based on this version of the geometry for simulation using `ATLAS-R2-2016-01-00-01`
as the two tags are the same except for the change from `PixelBarrelService-29` to `PixelBarrelService-30`.
The change here is "PST fwd region bug fix (z=3100 to 3200)".