TrigOnlineMonitor: move TrigOpMonitor configuration to runHLT
Move the `TrigOpMonitor` setup into the main runHLT job options. It is
really only needed for online running and it avoids having to check if
Luminosity and MagneticField CondAlgs are available which is difficult
to do in CA-based job options.

This also removes the last use-case of the post-exec in athenaHLT.

Relates to ATR-26800.