Read/Write/Decor handles for TruthCollectionMaker
Read/Write/Decor handles for TruthCollectionMaker

This MR provides read/write/decor handles for TruthCollectionMaker as part of the work to prepare the derivation framework for MT. Note that this is insufficient to make this tool wholly compliant with MT, because it also accesses the metadata store via a non-context-aware method:
https://gitlab.cern.ch/atlas/athena/-/blob/master/PhysicsAnalysis/DerivationFramework/DerivationFrameworkMCTruth/src/TruthCollectionMaker.cxx#L90-114

in order to acquire some metadata about whether the events were generator with Sherpa. The preferred means of addressing this will be to move this property to the config level to avoid having to use metadata handles in this tool. This will be addressed in a later MR.