Read/Write/Decor handles for TauTruthMatchingWrapper
Read/Write/Decor handles for TauTruthMatchingWrapper

This provides R/W/D handles for TauTruthMatchingWrapper as part of the effort to prepare the derivation framework for AthenaMT running. Although this is sufficient for this tool, it calls another tool (TauTruthMatchingTool) which does decoration without context-aware decorators. This will also have to be migrated.