SGTools Simplify DataStore::m_keyMap.
SGTools Simplify DataStore::m_keyMap.

Remove the index field from the payload of DataStore::m_keyMap.

Simplifying this in order to move towards removing locking from proxy_exact.

We were keeping track of the index in m_proxies of each sgkey in order to
speed up removeProxy.  However, the main use of removeProxy is from clearStore,
and in that case we already know the index because we're iterating over
m_proxies.

This does mean that direct calls to removeProxy will be a bit slower,
since now we do a linear search to find the proxy in m_proxies.
However, this should be something that happens only rarely.