TauAnalysisTools: add support for DeepSet Tau ID WP
TauAnalysisTools: add support for DeepSet Tau ID WP

Hello,

This MR is adding support for DeepSet Tau ID WPs in TauAnalysisTools.
We can now require taus to pass e.g. the medium DeepSet JetID WP by having a `JetIDWP: JETIDDEEPSETMEDIUM` in a TauAnalysisTools selection file.

The ability to cut on the DeepSet score (instead of the WP) will be added in a follow-up MR.

DeepSet Tau ID is available in 23.0.14 DAODs, try it out!

Cheers,
Bertrand