TrigT2CaloCommon: always use Ringer in fastCaloRecoSequence
Remove the option to enable/disable the Ringer in
`fastCaloRecoSequence` because in any case only one instance (with a
unique name) of `T2CaloEgammaReFastAlgo` is used. This means that even
if `fastCaloRecoSequence` is called once with and once without Ringer,
the first one wins. We were lucky that the physics sequences always got
configured first (with Ringer enabled) and then the setup of the
calibration sequence (without Ringer) just re-used the existing
algorithm instance.