AsgComponentConfig Re-Design, master branch (2023.02.21.)
Changed one of the internal variables of asg::AsgComponentConfig.

The type used so far is just not sitting well with ROOT I/O. So
rather than using an std::tuple type, switched to using a simple
struct instead.

Added a unit test that would test the type of I/O which would fail
with the previous implementation of asg::AsgComponentConfig, but
would succeed with this re-designed setup.