TileMonitoring: Update monitoring using FELIX and Legacy containers
TileDigitsFlxMonitorAlgorithm is modified to select desired set of 7(or 16) samples to compare with legacy digits. Introduced a new array to reject empty channels of a module in digits comparison.
The default value of 'last sample' is changed to 15 to use 0 to 15 samples.

TileRawChannelFlxMonitorAlgorithm has been updated to avoid comparison of empty channel.

Made the algorithm simpler by removing unwanted variables.