Allow the mu+tau tnp chains to be used in the ID Trigger analyses
Allow the mu+tau tnp chains to be used

Some of the tau Rois in the mu+tau tnp chains get "_probe" added to their name
which breaks all the usual feature retrieval. This hack allows the roi name
to be automatically mangled with the "_probe" extension in the case that no
rois are found.