AthenaConfiguration+TriggerJobOpts: auto-configure HLT conditions/geo tag
Move the conditions/geometry tag deduction for the HLT into the main
flags (`IOVDb.GlobalTag` and `GeoModel.AtlasVersion`). For the moment,
we only use a dedicated tag for data(-taking). For MC, we rely on the
default auto-configuration.

Also remove the now obsolete `Trigger.OnlineCondTag` and
`Trigger.OnlineGeoTag` flags and adjust the RDOtoRDOtrigger transform to
set the flags directly.

Relates to ATR-27052.