Muon trigger CA migration of MuonCombinedAlg
Muon trigger CA migration of MuonCombinedAlg

Using the CA based config of the MuonCOmbinedAlg in the trigger (ATR-25086). Updated a few parameters of teh track builder/fitter to bring in line with the legacy configuration. As the CA based config is more correct for some tools that we share between SA and combined muons in the trigger (in the legacy config we needed to use the MuonTrackSumaryTool rather than the CombinedMuonTrackSummaryTool in a couple of places also for combined muons, but can now use the proper versions). As such, the HypoAlg has been updatedt o compare the pre-extrapolated MS tracks when determining if muons between different steps are the same, rather than the extrapolated tracks which can be different for the MS-only and combined muon steps.