xAODEventInfoCnv: Do not copy detector flags in EventInfoOverlay.
xAODEventInfoCnv: Do not copy detector flags in EventInfoOverlay.

EventInfoOverlay was copying all the detector flags from the background file.
The problem is that means that when we get to the reco step, these flags
will be locked.  If reco then tries to set any different flags,
it will fail with an error.

So don't copy the detector flags (except for Core, which  was already
being done explicitly).  Reco will set the flags again if appropriate.

For now, do this only for simulation.