Fix of missing event generator weight names from the DAOD_TRUTHx formats
Fix of missing event generator weight names from the DAOD_TRUTHx formats

ATLASG-2475

This (probably temporary) fix applies the IOVDbMetaDataTool in the EVNT->DAOD_TRUTHx workflow. Without this adjustment the event generator weight names are not stored and the formats are useless for many analyses. The issue and the reason the fix works are rather hazily understood so this may be revisited during a wider clean-up of this metadata, but for now the samples are needed for physics studies.