Remove MuonRecFlags/MuonCombinedRecFlags from the trigger
Remove MuonRecFlags/MuonCombinedRecFlags from the trigger

Removing the setting of MuonRecFlags and MuonCombinedRecFlags from the trigger as all the correponding config flags are set correctly and the old flags are no longer used. Also removing the DisableMdtT-Fit modifier as the CA based configuration sets is correct in all use cases without this, so it isn't needed anymore.