Adding additional filter outputs to PHYS/LITE
Adding additional filter outputs to PHYS/LITE

In !61340 we added an additional filter that analyses like to have to
DerivationFrameworkMCTruth and to the output of the TRUTH derivations.
This merge request adds the filter output to the PHYS and PHYSLITE
derivations as well. It's 3 extra floats per event decorated onto the
EventInfo (I took all the filter outputs, while we're here).