Patches to running TRUTH derivation on HITS input
Patches to running TRUTH derivation on HITS input

Two little tweaks were needed to run TRUTH derivations on HITS input:

- One of the CA calls was missing the flags, just a simple typo

- The metadata does not already exist in a HITS file (like it does in an
AOD), so it has to be created as though we are running on an EVNT file.

With these two small changes, the TRUTH derivations run on HITS input
without a problem.