Allow NSW lookup tables to have duplicated detector elements for the split robs
Allow NSW lookup tables to have duplicated detector elements for the split robs

In the NSW apparently eacvhj detector element will be read out by (up to) 4 ROBS
and as such multiple ROBIDs are required for the detector elements.

As a quick hack the detector elements are added multiple times, once for each ROBID,
so this change should removed the duplicate hash IDs from the output list for the
MM And sTGC instances. See !2

A more complete fix might be required in the near future.