Fixing 'Overwriting primary component' warnings in TRUTH[0,1,3]
Fixing 'Overwriting primary component' warnings in TRUTH[0,1,3]

ATLASG-2480

This fixes three 'Overwriting primary component' warnings in the output of TRUTH[0,1,3]. They were caused by several tools being declared as primary to the same component accumulator. Since none of them were needed in a later getPrimaryAndMerge step none of them need to primary so this MR rectifies that and fixes the warning message.

At the same time an unrelated default property is changed from a (silly) placeholder value to the value that is most commonly used.

Neither change will affect the content of TRUTH[0,1,3]