TrigTauHypo: add property to count tracks above pt threshold in the perf step
TrigTauHypo: add property to count tracks above pt threshold in the perf step

Hello,

This MR is adding a property to the main tau hypo tool, to count tracks above a custom pt threshold in the perf stage.
Currently, a 0.8 GeV threshold is used for HLT tau track reconstruction, but the tau track multiplicity (i.e. in `tau->nTracks()` and `tau->nTracksIsolation()`) is based on tracks with pt > 1 GeV, as set in the [InDetTrackSelectorTool](https://gitlab.cern.ch/atlas/athena/-/blob/master/Trigger/TrigAlgorithms/TrigTauRec/python/TrigTauToolsConfig.py#L48-64).
In the 'idperf->perf' HLT tau selection, we require `tau->nTracks()<=3` and `tau->nTracksIsolation()<1`, and we have evidence this causes inefficiency at high mu because we tend to include soft pileup tracks when counting tracks.
Raising the track pt threshold to 1.5 GeV (just in the final hypo when counting tracks) increases the track counting efficiency for true taus by about 0.5% for 1p and 2% 3p, and the rates are under control!

For now, this MR only introduces the property for validation, but it is not in use (set to 0). The anticipated value to be set after validation is 1.5 GeV.

Adding the 'urgent' flag, as we need to validate this urgently for the 2023 startup.

Cheers,
Bertrand