Migration of tag and probe chains to CA
Migration of tag and probe chains to CA

An attempt to get the probe legs of tag and probe chains working in CA configuration (ATR-27099). Have setup the RoI tools and any specified input maker algs supplied to InViewRecoCA to be copied with the updated RoI collection names for probe legs, and have made sure the default view names are set accordingly for probe legs. Then also skipped the cloning we usually do for the legacy MenuSequence assuming MenuSequenceCA which should set isProbe in the InViewRecoCA directly.