TrigTauRec and TrigTauHypo: count tracks above 1.5 GeV in perf, adjust DeepSet ID WP
TrigTauRec and TrigTauHypo: count tracks above 1.5 GeV in perf, adjust DeepSet ID WP

Hello,

This MR is a follow up of !61823 . To determine whether a HLT tau candidate passes the "perf_tracktwoMVA" selection, we now count tracks above pt=1.5 GeV instead of 1 GeV, and we require at most 3 such tracks in the core region of the tau (R<0.2), and at most 1 track in the isolation region (0.2<R<0.4).
The DeepSet ID WPs are slightly tightened for multiprong taus to keep rates at a comparable level. The pt threshold above which no selection is applied to HLT taus is raised to 430 GeV to slightly decrease the rates, it had recently been lowered from 440 to 420 GeV (part of the rate increase is harmless as these events are also triggered by HLT_j420, but we couldn't evaluate the increase in unique rate).

The motivations and performance studies are detailed in [these slides](https://indico.cern.ch/event/1269581/contributions/5332331/attachments/2618792/4527589/Martin_TauTrig_27Mar23_new.pdf).

Adding the 'urgent' flag as we need this for the 2023 startup.

Cheers,
Bertrand