Hopefully circumvent default Roi Z width change for the LRT lepton chains
Hopefully circumvent default Roi Z width change for the LRT lepton chains

The Roi Z width change has generally been successful for the tracking from the beamline,
but has had an unexpected impact on the LRT lepton efficiency due to the extrapolation
of large radius tracks to the beam line having a larger spread in z that the beamspot.

This is well known, and the LRT already includes a wider internal selection of +- 500 mm
for this very purpose. However there is still a reduced efficiency and it is suspected
that limitations on the way candidates are selected within the eta limits are not optimal
for such a configuration, and as such despite the wider internal z range used some tracking
efficiency remains.

This change will hopefully reinstate the |z|<225 mm selection for the L=RT lepton
signatures.

This is in a sense a bug fix, since the Roi updater had no option to allow setting
of the z width, which it should have done (in fact all the Roi updaters should) so this
is really a fix for a missing featurethat should have been included already to allow this
change just from the python configuration.

This change is expected to change the trigger counts for the LRT chains, to bring
them in line with what they were before the default z width change.