Update list of event views in EDM for muon triggers
Update list of event views in EDM for muon triggers

Updating the list of event view names for SA muon container in the EDM. This is needed because of the CA migration, the views are named automatically, and the former naming didn't follow a similar enough naming convention to continue to be used.

Bacause the L2 SA step is not completely migrated to CA, the EDM updates cause soem problems with the purely new JO trigger test, so temporarily renaming the views used in the newJO test case until the step is fully CA based (this way, the EDM won't need to be changed again when the migration for this step is completed).

Fixes ATR-27231.