AthenaKernel+CxxUtils: Mixed container trimming workaround.
CxxUtils: Add an option to ConcurrentRangeMap::trim to allow removing all elements.

By default, trim() would not remove the last element from a container.
Add an option to allow the last element to be removed as well.
To be used to work around inconsistent cleaning of tracking DD objects.
See ATLASRECTS-7421.