TrackingGeometryCondAlg: use configurable for EnvelopeDefinitionSvc property
Set the `EnvelopeDefinitionSvc` property to the Configurable instance
instead of the service name string. In the Muon case, remove the
conditional setting of this property (and some other) to bring it in
line with the other geometry builders.

This remove thousands(!) of warnings of this type in (legacy) trigger
jobs:

Py:conf2toConfigurable WARNING  Component: 'AtlasGeometry_EnvelopeDefSvc' is of type string, no conversion, some properties possibly not set?