TileMonitoring: Decoration fix.
TileMonitoring: Decoration fix.

A DecorHandle should only be applied to elements of the container with
which the handle was declared.  This implies in particular that to attach
decorations to a temporary object that is not recorded, one should not
use a handle, but instead just a Decorator.

Set SuppressOutputDependence for EventCleaningTool.

See !62000.