Add b-jet preselction and tau tag and probe chains to the tier0 monitoring
Add b-jet preselction and tau tag and probe chains to the tier0 monitoring

For some reason the b-jet presection tracking was not yet included in the tiet0.
The tau tag an d probe also needs to be added. There may be issues with the correct
function of the tau and and probe, but these will be addressed in a subsequent
merge request.