preInclude.SleptonsLLP.py: Drop use of ExtraParticles.PDGHelpers module
preInclude.SleptonsLLP.py: Drop use of ExtraParticles.PDGHelpers module - not required for digitization

As this branch is not intended to be used for simulation (only digitization onwards). Then it is safe
to drop the usage of the `ExtraParticles.PDGHelpers` module, which is intended to manage files being
updated at multiple points during the configuration of the the simulation step - this functionality
is not required for later steps. This should fix the issues seen in ATLMCPROD-10493 (ADCSUPPORT-5537).