Add AtlasFieldCacheCondAlg to L1 muon simulation config
Add AtlasFieldCacheCondAlg to L1 muon simulation config

Since the muon RDO to digit algorithm needs the objects created by the AtlasFieldCacheCondAlg, adding this to the configuration used in the L1 muon simulation (otherwise it relies on the HLT muon configuration to be present).

This will allow to run non-muon HLT chains alone in the new JO tests.