Fix test_Digi_tf_mc20e_MP_presampling_reproducibility.sh (ATLASSIM-6217)
This ART job compares the output between Serial Athena and AthenaMP
digitization jobs. The output agrees already apart from the GenParticle
masses. The TP conversion for these variables is lossy, but reproducible,
so the fix is to ensure that the files being compared have gone through
the same number of TP conversion cycles. Currently the AthenaMP
jobs have an extra cycle due to the RDOMerge step to combine the outputs
of the various workers. Adding a dummy RDOMerge step for the Serial
Athena job should fix the differences.