fix mu+tau tnp navigation access for the tier0
fix mu+tau tnp navigation access for the tier0

The tag and probe chains mangle the names of the Rois used in the mu+tau tan
and probe chains so that they are different from the same sequence but not run
as a tag and probe chain.

This fix, searches *also* for the mangled named Rois for the ti9er 0 analyses.