TriggerJobOpts: make runHLT the main CA config module
The `runHLT.py` module is now the main CA config module. It can be used
in `athena` and `athenaHLT` via different entry points.

It replaces `runHLT_standalone_newJO.py` which is kept as a wrapper
until all references to it have been changed.

TODO: does not work yet in athenaHLT