Replace the use of the pointer to Trk::Track in the TauTrack access
Replace the use of the pointer to Trk::Track in the TauTrack access

The "TauTrack" particles in the offline tau, are not actual tracks but some limited custom class that
inherits from IParticle. To access the actual track one has to use the track()
method of the class to return the pointer to the Trk::Track.

However, this doesn;t work in the derivations so instead this has
to be replaced by the tedious to use readLinks() method which returns
a vector of ElementLinks and then one has to loop through there for
the single track.