ElectronPhotonSelectorTools: Add option to use avg mu in el LH tool
ElectronPhotonSelectorTools: Add option to use avg mu in el LH tool

Normally one has to pass the mu to the accept function.
The goal of this merge is to have an option to use it automatically
via the nominal `accept(IParticle*)` function.

Does not affect existing functionality, only adds new option `useAverageMu`.
Intended mainly for analysis of online electrons, where one needs to use mu
but the tool is called via `AsgSelectionAlg`.