ParticleGun: fix pdgid and status of incoming particles
ParticleGun: fix pdgid and status of incoming particles

Hello,

In the definition of ad-hoc beam particles, it seems that the PDG ID and status have been swapped.
Concerning the ad-hoc beam energy, that seems a bit overkill to grab it from runArgs, so I didn't dare touch this.
This MR should have no consequence for the generated final state particles of interest.

Cheers,
Bertrand