Avoid removing particles from the GenEvent after SimHit creation has started
Add TruthPreselectionTool

This removes the decay chains of quasi-stable particles from the
`GenEvent` and adds the "ShadowParticleId" `HepMC3::IntAttribute`
to all `GenParticle` objects in the filtered `GenEvent` which
should be passed to the simulation in the future (including
`identified quasi-stable particles`) to provide a link to the
corresponding particles in the original (unfiltered) `GenEvent`.

Next step: consider moving the Tool inside the InputConverter?