trigbs_extractStream.py: Add the possibility to write events with different stream tags to the same output bytestream file.
trigbs_extractStream.py: Add the possibility to write events with different stream tags
to the same output bytestream file.

This functionality will be used in the trigger transform for the  debug stream recovery
where accepted events from different streams will be collected in one file for processing
in Tier-0.

Also relevant for: ATR-26595